+++ .eslintignore
@@ -1,2 +1,2 @@
  
-jobs/csvjson/*.js
+jobs/core/csvjson/*.js
