+++ .eslintrc.js
@@ -1,18 +1,14 @@
 module.exports = {
   env: {
     commonjs: true,
-    es6: true,
+    es2020: true,
     node: true,
   },
   extends: [
     'airbnb-base',
   ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
   parserOptions: {
-    ecmaVersion: 2018,
+    ecmaVersion: 11,
   },
   rules: {
   },
