+++ .github/FUNDING.yml
@@ -1,4 +1,4 @@
 # These are supported funding model platforms
 
 ko_fi:  eusoumaxi
-custom: https://www.buymeacoffee.com/eusoumaxi
+custom: https://paypal.me/covid19server
