+++ .gitignore
@@ -84,3 +84,4 @@ dist/
 package-lock.json
 jobs/db/
 ./jobs/db/
+./store/
