+++ jobs/core/helper/csvjson/index.js
@@ -1,0 +1,4 @@
+var app 		= require('./src/app');
+var stream 		= require('./src/stream');
+app.stream 		= stream;
+module.exports 	= app;
