+++ jobs/core/index.js
@@ -1,0 +1,4 @@
+const { dataCore } = require('./getCountryCore');
+const { getTimeLine } = require('./getTimeLineCore');
+
+module.exports = { dataCore, getTimeLine };
