+++ jobs/services/index.js
@@ -1,0 +1,10 @@
+// services
+
+// Country
+const { countryVenezuela } = require('./country/venezuela');
+
+// timeLine
+const { timeLineColombia } = require('./timeline/colombia');
+const { timeLineVenezuela } = require('./timeline/venezuela');
+
+module.exports = { countryVenezuela, timeLineColombia, timeLineVenezuela };
