+++ src/routes/api.js
@@ -1,10 +1,13 @@
-/* eslint-disable no-unused-expressions */
 const { Router } = require('express');
 const allJson = require('../utils/data/all');
 const DataServices = require('../services/getData');
 const cacheResponse = require('../utils/cache');
 const { threeHour } = require('../utils/time');
-const { dataFilterHelp } = require('../utils/helper/servicesHelper');
+const {
+  dataFilterHelp,
+  getCountriesURL,
+  error404Countries,
+} = require('../utils/helper/servicesHelper');
 
 const dataService = new DataServices();
 const router = Router();
@@ -21,7 +24,6 @@ router.get('/', async (_req, res, next) => {
 router.get('/summary', async (_req, res, next) => {
   cacheResponse(res, threeHour);
   try {
-    // throw new Error('This iss an error');
     const dataSevices = await dataService.getSummaries();
     res.status(200).json(dataSevices);
   } catch (err) {
@@ -30,6 +32,7 @@ router.get('/summary', async (_req, res, next) => {
 });
 router.get('/country', async (_req, res, next) => {
   cacheResponse(res, threeHour);
+
   try {
     const dataSevices = await dataService.getDataCountries();
     res.status(200).json(dataSevices);
@@ -40,9 +43,8 @@ router.get('/country', async (_req, res, next) => {
 router.get('/all', async (_req, res, next) => {
   cacheResponse(res, threeHour);
   try {
-    const dataSevicesCountry = await dataService.getTimelineAll();
-    const dataSevicesTimeLine = await dataService.getTimelineAll();
-    res.status(200).json({ dataSevicesCountry, dataSevicesTimeLine });
+    const dataSevicesCountry = await dataService.getDataAllCountryData();
+    res.status(200).json(dataSevicesCountry);
   } catch (err) {
     next(err);
   }
@@ -60,11 +62,14 @@ router.get('/timeline', async (_req, res, next) => {
 router.get('/country/:countries', async (req, res, next) => {
   cacheResponse(res, threeHour);
   const { countries } = req.params;
+  const country = getCountriesURL(countries);
   try {
-    const data = await dataService.getCountries(countries);
-    res.status(200).json({
-      ...data,
-    });
+    if (country === null) {
+      res.status(404).json(error404Countries);
+    } else {
+      const data = await dataService.getCountries(country);
+      res.status(200).json({ ...data });
+    }
   } catch (err) {
     next(err);
   }
@@ -72,9 +77,14 @@ router.get('/country/:countries', async (req, res, next) => {
 router.get('/country/:countries/:statep', async (req, res, next) => {
   cacheResponse(res, threeHour);
   const { countries, statep } = req.params;
+  const country = getCountriesURL(countries);
   try {
-    const data = await dataService.getState(countries, statep);
-    res.status(200).json(data);
+    if (country === null) {
+      res.status(404).json(error404Countries);
+    } else {
+      const data = await dataService.getState(country, statep);
+      res.status(200).json(data);
+    }
   } catch (err) {
     next(err);
   }
@@ -82,9 +92,14 @@ router.get('/country/:countries/:statep', async (req, res, next) => {
 router.get('/country/:countries/:statep/:cityp', async (req, res, next) => {
   cacheResponse(res, threeHour);
   const { countries, statep, cityp } = req.params;
+  const country = getCountriesURL(countries);
   try {
-    const data = await dataService.getCity(countries, statep, cityp);
-    res.status(200).json(data);
+    if (country === null) {
+      res.status(404).json(error404Countries);
+    } else {
+      const data = await dataService.getCity(country, statep, cityp);
+      res.status(200).json(data);
+    }
   } catch (err) {
     next(err);
   }
@@ -92,9 +107,14 @@ router.get('/country/:countries/:statep/:cityp', async (req, res, next) => {
 router.get('/timeline/:countries', async (req, res, next) => {
   cacheResponse(res, threeHour);
   const { countries } = req.params;
+  const country = getCountriesURL(countries);
   try {
-    const data = await dataService.getTimeLine(countries);
-    res.status(200).json(data);
+    if (country === null) {
+      res.status(404).json(error404Countries);
+    } else {
+      const data = await dataService.getTimeLine(country);
+      res.status(200).json(data);
+    }
   } catch (err) {
     next(err);
   }
@@ -102,23 +122,36 @@ router.get('/timeline/:countries', async (req, res, next) => {
 router.get('/timeline/:countries/provinces', async (req, res, next) => {
   cacheResponse(res, threeHour);
   const { countries, statep } = req.params;
+  const country = getCountriesURL(countries);
   try {
-    const data = await dataService.getTimeLineInfo(countries, statep);
-    res.status(200).json(data);
+    if (country === null) {
+      res.status(404).json(error404Countries);
+    } else {
+      const data = await dataService.getTimeLineInfo(country, statep);
+      res.status(200).json(data);
+    }
   } catch (err) {
     next(err);
   }
 });
-router.get('/timeline/:countries/provinces/:statep/', async (req, res, next) => {
-  cacheResponse(res, threeHour);
-  const { countries, statep } = req.params;
-  try {
-    const data = await dataService.getTimeLineCity(countries, statep);
-    res.status(200).json(data);
-  } catch (err) {
-    next(err);
+router.get(
+  '/timeline/:countries/provinces/:statep/',
+  async (req, res, next) => {
+    cacheResponse(res, threeHour);
+    const { countries, statep } = req.params;
+    const country = getCountriesURL(countries);
+    try {
+      if (country === null) {
+        res.status(404).json(error404Countries);
+      } else {
+        const data = await dataService.getTimeLineCity(country, statep);
+        res.status(200).json(data);
+      }
+    } catch (err) {
+      next(err);
+    }
   }
-});
+);
 router.get('/filters', async (req, res, next) => {
   const { date, endDate, country } = req.query;
 
@@ -127,19 +160,26 @@ router.get('/filters', async (req, res, next) => {
       message: 'There was an error, check the data',
       country: country || 'country is required',
       date: date || 'date is required',
-      endDate: endDate
+      endDate: endDate,
     },
     dataFilterHelp,
   ];
+
   try {
     if (!country) {
       res.status(400).json(errorData);
+    } else if (getCountriesURL(country) === null) {
+      res.status(404).json(error404Countries);
     } else if (!Date.parse(date)) {
       res.status(400).json(errorData);
     } else if (endDate && !Date.parse(endDate)) {
       res.status(400).json(errorData);
     } else {
-      const data = await dataService.filters(country, date, endDate);
+      const data = await dataService.filters(
+        getCountriesURL(country),
+        date,
+        endDate
+      );
       res.status(200).json(data);
     }
   } catch (err) {
