+++ src/utils/cache.js
@@ -1,4 +1,4 @@
-const { config } = require('../config');
+const { config } = require('../../config');
 
 function cacheResponse(res, seconds) {
   if (!config.dev) {
